refactor: composable useGroupTooltip
